Swap bogus fedId types with correct Long types

+++ src/main/scala/model/competition/LeagueDao.scala
@@ -3,5 +3,5 @@ package main.scala.model.competition
 import main.scala.model.generic.GenericDao
 
 trait LeagueDao extends GenericDao[League, java.lang.Long] {
-  def findBySlug(fedId: Int, slug: String): Option[League]
+  def findBySlug(fedId: Long, slug: String): Option[League]
 }
